requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
PyMuPDF
pytesseract
waitress